+++ FastLog.Net/Helpers/ExtendedMethods/DateTimeExtendedMethods.cs
@@ -24,10 +24,14 @@ namespace FastLog.Helpers.ExtendedMethods
     /// </summary>
     internal static class DateTimeExtendedMethods
     {
-        public static string ToFriendlyDateTime(this DateTime dateTime) => $"{ToFriendlyDate(dateTime)} {ToFriendlyTime(dateTime)}";
+        public static string ToFriendlyDateTime(this DateTime dateTime, bool ToUTC = false) => $"{ToFriendlyDate(dateTime,ToUTC)} {ToFriendlyTime(dateTime, ToUTC)}";
 
-        public static string ToFriendlyDate(this DateTime dateTime) => dateTime.ToString("yyyy-MM-dd");
+        public static string ToFriendlyDate(this DateTime dateTime, bool ToUTC = false) => ToUTC ? dateTime.ToString("yyyy-MM-dd") : dateTime.ToUniversalTime().ToString("yyyy-MM-dd");
 
-        public static string ToFriendlyTime(this DateTime dateTime) => dateTime.ToString("HH:mm:ss");
+        public static string ToFriendlyTime(this DateTime dateTime, bool ToUTC = false) => ToUTC ? dateTime.ToString("HH:mm:ss") : dateTime.ToUniversalTime().ToString("HH:mm:ss");
+
+
+
+
     }
 }