+++ FastLog.Net/Helpers/FileHelper.cs
@@ -18,19 +18,6 @@ using System.IO;
 
 namespace FastLog.Helpers
 {
-
-    //#Refactor : In codes below we used "ReaderWriterLockSlim" for thread-safety file access , BUT !
-    // Of course there is better solution to achive that.
-
-    //Known Issue : "ReaderWriterLockSlim" is thread-safe but could not be used in async method and also is not process safe.
-    // See accepted answer of this thread for more info :
-    // https://stackoverflow.com/questions/19659387/readerwriterlockslim-and-async-await
-
-
-    /// <summary>
-    /// Thread-Safe File Access ( read and write ) with "ReaderWriterLockSlim" class.
-    /// Note : "ReaderWriterLockSlim" is thread-safe but is NOT process safe.
-    /// </summary>
     internal static class FileHelper
     {