+++ CHANGELOG.md
@@ -14,4 +14,8 @@
 ### 1.1.8550 (2023-05-30)
   * Improve Internal Logger Thread-Safety.
   * Add Nuget package address to the file header.
+  * Improve FastLog.Net engine.
+  * Improve Test App.
+  * Add UTC Date Time to the file header.
+  * Improve FastLog.Net Engine exceptions handling.